# docs/OpenApi.yaml
openapi: 3.1.0
info:
  title: Sistema de Gestión Ganadera - API Docs
  version: 0.0.1
  description: >-
    Documentación interactiva de la API REST para la gestión y trazabilidad
    ganadera de la empresa. Permite a los desarrolladores explorar y probar los
    endpoints de la API, facilitando la integración con sistemas externos y
    aplicaciones móviles.
servers:
  - url: 'http://localhost:8000/api'
...
security:
  - http: []
paths:
...
  '/v1/animales/{id}':
    get:
      operationId: animales.show
      description: >-
        Retorna la información completa de un animal específico por su ID,

        incluyendo sus relaciones con raza, potrero, historial de pesajes y
        tratamientos.
      summary: Obtener detalle de un animal
      tags:
        - Animal
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador único del animal.
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    const: Animal obtenido correctamente.
                  data:
                    anyOf:
                      - type: string
                      - $ref: '#/components/schemas/AnimalResource'
                required:
                  - message
                  - data
      security: []
    put:
      operationId: animales.update
      description: >-
        Actualiza la información de un animal existente por su ID.

        Si se modifica la asignación de potrero, valida que el nuevo potrero
        tenga capacidad disponible.
      summary: Actualizar datos de un animal
      tags:
        - Animal
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador único del animal a actualizar.
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAnimalRequest'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    const: Animal actualizado correctamente.
                  data:
                    anyOf:
                      - type: string
                      - $ref: '#/components/schemas/AnimalResource'
                required:
                  - message
                  - data
        '422':
          $ref: '#/components/responses/ValidationException'
      security: []
    delete:
      operationId: animales.destroy
      description: >-
        Elimina el registro de un animal del sistema por su ID.

        Impide la eliminación si el animal posee registros de pesaje en su
        historial.
      summary: Eliminar un animal
      tags:
        - Animal
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador único del animal a eliminar.
          schema:
            type:
              - integer
              - string
      responses:
        '204':
          description: No content
      security: []
...
  '/v1/categorias/{id}':
    get:
      operationId: categorias.show
      description: Retorna la información de una categoría específica por su ID.
      summary: Obtener detalle de una categoría
      tags:
        - Categoria
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    const: Categoría obtenida correctamente.
                  data:
                    anyOf:
                      - type: string
                      - $ref: '#/components/schemas/CategoriaResource'
                required:
                  - message
                  - data
      security: []
    put:
      operationId: categorias.update
      description: Actualiza la información de una categoría existente por su ID.
      summary: Actualizar una categoría
      tags:
        - Categoria
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoriaRequest'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    const: Categoría actualizada correctamente.
                  data:
                    anyOf:
                      - type: string
                      - $ref: '#/components/schemas/CategoriaResource'
                required:
                  - message
                  - data
        '422':
          $ref: '#/components/responses/ValidationException'
      security: []
    delete:
      operationId: categorias.destroy
      description: Elimina del sistema la categoría indicada por su ID.
      summary: Eliminar una categoría
      tags:
        - Categoria
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No content
      security: []
...
  '/v1/insumos/{id}':
    get:
      operationId: insumos.show
      description: Retorna la información de un insumo específico por su ID.
      summary: Obtener detalle de un insumo
      tags:
        - Insumo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    const: Insumo obtenido correctamente.
                  data:
                    anyOf:
                      - type: string
                      - $ref: '#/components/schemas/InsumoResource'
                required:
                  - message
                  - data
      security: []
    put:
      operationId: insumos.update
      description: Actualiza la información y existencias de un insumo existente por su ID.
      summary: Actualizar datos de un insumo
      tags:
        - Insumo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInsumoRequest'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    const: Insumo actualizado correctamente.
                  data:
                    anyOf:
                      - type: string
                      - $ref: '#/components/schemas/InsumoResource'
                required:
                  - message
                  - data
        '422':
          $ref: '#/components/responses/ValidationException'
      security: []
    delete:
      operationId: insumos.destroy
      description: Elimina el insumo indicado del sistema por su ID.
      summary: Eliminar un insumo
      tags:
        - Insumo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No content
      security: []
...
  '/v1/pesajes/{id}':
    get:
      operationId: pesajes.show
      description: Retorna la información completa de un registro de pesaje por su ID.
      summary: Obtener detalle de un pesaje
      tags:
        - Pesaje
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    const: Pesaje obtenido correctamente.
                  data:
                    anyOf:
                      - type: string
                      - $ref: '#/components/schemas/PesajeResource'
                required:
                  - message
                  - data
      security: []
    put:
      operationId: pesajes.update
      description: Actualiza los datos de un registro de pesaje existente por su ID.
      summary: Actualizar un pesaje
      tags:
        - Pesaje
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePesajeRequest'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    const: Pesaje actualizado correctamente.
                  data:
                    anyOf:
                      - type: string
                      - $ref: '#/components/schemas/PesajeResource'
                required:
                  - message
                  - data
        '422':
          $ref: '#/components/responses/ValidationException'
      security: []
    delete:
      operationId: pesajes.destroy
      description: Elimina el registro de pesaje indicado del sistema.
      summary: Eliminar un pesaje
      tags:
        - Pesaje
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No content
      security: []
...
  '/v1/potreros/{id}':
    get:
      operationId: potreros.show
      description: Retorna la información de un potrero específico por su ID.
      summary: Obtener detalle de un potrero
      tags:
        - Potrero
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: integer
                const: 200
      security: []
    put:
      operationId: potreros.update
      description: Actualiza la información de un potrero existente.
      summary: Actualizar datos de un potrero
      tags:
        - Potrero
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePotreroRequest'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: integer
                const: 200
        '422':
          $ref: '#/components/responses/ValidationException'
      security: []
    delete:
      operationId: potreros.destroy
      description: >-
        Elimina el potrero indicado. Falla si el potrero tiene animales
        asignados actualmente (RN-02).
      summary: Eliminar un potrero
      tags:
        - Potrero
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: integer
                const: 200
      security: []
...
  '/v1/razas/{id}':
    get:
      operationId: razas.show
      description: Retorna la información de una raza específica por su ID.
      summary: Obtener detalle de una raza
      tags:
        - Raza
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    const: Raza obtenida correctamente.
                  data:
                    anyOf:
                      - type: string
                      - $ref: '#/components/schemas/RazaResource'
                required:
                  - message
                  - data
      security: []
    put:
      operationId: razas.update
      description: Actualiza la información de una raza existente por su ID.
      summary: Actualizar una raza
      tags:
        - Raza
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRazaRequest'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    const: Raza actualizada correctamente.
                  data:
                    anyOf:
                      - type: string
                      - $ref: '#/components/schemas/RazaResource'
                required:
                  - message
                  - data
        '422':
          $ref: '#/components/responses/ValidationException'
      security: []
    delete:
      operationId: razas.destroy
      description: Elimina del sistema la raza indicada por su ID.
      summary: Eliminar una raza
      tags:
        - Raza
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No content
      security: []
...
  '/v1/tratamientos/{id}':
    get:
      operationId: tratamientos.show
      description: Retorna la información de un tratamiento específico por su ID.
      summary: Obtener detalle de un tratamiento
      tags:
        - Tratamiento
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    const: Tratamiento obtenido correctamente.
                  data:
                    anyOf:
                      - type: string
                      - $ref: '#/components/schemas/TratamientoResource'
                required:
                  - message
                  - data
      security: []
    put:
      operationId: tratamientos.update
      description: Actualiza la información de un tratamiento existente por su ID.
      summary: Actualizar un tratamiento
      tags:
        - Tratamiento
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTratamientoRequest'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    const: Tratamiento actualizado correctamente.
                  data:
                    anyOf:
                      - type: string
                      - $ref: '#/components/schemas/TratamientoResource'
                required:
                  - message
                  - data
        '422':
          $ref: '#/components/responses/ValidationException'
      security: []
    delete:
      operationId: tratamientos.destroy
      description: Elimina del catálogo el tratamiento indicado por su ID.
      summary: Eliminar un tratamiento
      tags:
        - Tratamiento
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No content
      security: []
...
  '/v1/tratamientos-aplicaciones/{id}':
    get:
      operationId: tratamientos-aplicaciones.show
      description: >-
        Retorna la información detallada de una aplicación de tratamiento por su
        ID.
      summary: Obtener detalle de una aplicación de tratamiento
      tags:
        - TratamientoAnimal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    const: Aplicación de tratamiento obtenida correctamente.
                  data:
                    anyOf:
                      - type: string
                      - $ref: '#/components/schemas/TratamientoAnimalResource'
                required:
                  - message
                  - data
      security: []
    put:
      operationId: tratamientos-aplicaciones.update
      description: >-
        Actualiza la información de una aplicación de tratamiento existente por
        su ID.
      summary: Actualizar aplicación de tratamiento
      tags:
        - TratamientoAnimal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTratamientoAnimalRequest'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    const: Aplicación de tratamiento actualizada correctamente.
                  data:
                    anyOf:
                      - type: string
                      - $ref: '#/components/schemas/TratamientoAnimalResource'
                required:
                  - message
                  - data
        '422':
          $ref: '#/components/responses/ValidationException'
      security: []
    delete:
      operationId: tratamientos-aplicaciones.destroy
      description: >-
        Elimina del sistema el registro de la aplicación de tratamiento
        indicada.
      summary: Eliminar aplicación de tratamiento
      tags:
        - TratamientoAnimal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No content
      security: []
...
  '/v1/unidades-medida/{id}':
    get:
      operationId: unidades-medida.show
      description: Retorna la información de una unidad de medida específica por su ID.
      summary: Obtener detalle de una unidad de medida
      tags:
        - UnidadMedida
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    const: Unidad de medida obtenida correctamente.
                  data:
                    anyOf:
                      - type: string
                      - $ref: '#/components/schemas/UnidadMedidaResource'
                required:
                  - message
                  - data
      security: []
    put:
      operationId: unidades-medida.update
      description: Actualiza la información de una unidad de medida existente por su ID.
      summary: Actualizar una unidad de medida
      tags:
        - UnidadMedida
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUnidadMedidaRequest'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    const: Unidad de medida actualizada correctamente.
                  data:
                    anyOf:
                      - type: string
                      - $ref: '#/components/schemas/UnidadMedidaResource'
                required:
                  - message
                  - data
        '422':
          $ref: '#/components/responses/ValidationException'
      security: []
    delete:
      operationId: unidades-medida.destroy
      description: Elimina del sistema la unidad de medida indicada por su ID.
      summary: Eliminar una unidad de medida
      tags:
        - UnidadMedida
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No content
      security: []
components:
  securitySchemes:
    http:
      type: http
      scheme: bearer
  schemas:
...
    AnimalResource:
      type: object
      properties:
        id_animal:
          type: integer
        numero_arete:
          type: string
        sexo:
          type: string
        fecha_nacimiento:
          type:
            - string
            - 'null'
        estado:
          type: string
        raza_id:
          type: integer
        potrero_id:
          type: integer
        raza:
          $ref: '#/components/schemas/RazaResource'
        potrero:
          $ref: '#/components/schemas/PotreroResource'
        pesajes:
          type: array
          items:
            $ref: '#/components/schemas/PesajeResource'
        tratamientos:
          type: array
          items:
            $ref: '#/components/schemas/TratamientoResource'
        created_at:
          type:
            - string
            - 'null'
        updated_at:
          type:
            - string
            - 'null'
      required:
        - id_animal
        - numero_arete
        - sexo
        - fecha_nacimiento
        - estado
        - raza_id
        - potrero_id
        - created_at
        - updated_at
      title: AnimalResource
    CategoriaResource:
      type: object
      properties:
        categoria_id:
          type: integer
        nombre:
          type: string
        tipo:
          type:
            - string
            - 'null'
        created_at:
          type:
            - string
            - 'null'
        updated_at:
          type:
            - string
            - 'null'
      required:
        - categoria_id
        - nombre
        - tipo
        - created_at
        - updated_at
      title: CategoriaResource
    InsumoResource:
      type: object
      properties:
        insumo_id:
          type: integer
        nombre:
          type: string
        categoria_id:
          type: integer
        precio:
          type: number
        unidad_medida_id:
          type: integer
        categoria:
          $ref: '#/components/schemas/CategoriaResource'
        unidad_medida:
          $ref: '#/components/schemas/UnidadMedidaResource'
        created_at:
          type:
            - string
            - 'null'
        updated_at:
          type:
            - string
            - 'null'
      required:
        - insumo_id
        - nombre
        - categoria_id
        - precio
        - unidad_medida_id
        - created_at
        - updated_at
      title: InsumoResource
...
    PesajeResource:
      type: object
      properties:
        pesaje_id:
          type: integer
        id_animal:
          type: integer
        peso_kg:
          type: number
        fecha_pesaje:
          type:
            - string
            - 'null'
        observaciones:
          type:
            - string
            - 'null'
        animal:
          type: object
          description: 'Resumen del animal, solo si la relación fue cargada.'
          properties:
            id_animal:
              type: integer
            numero_arete:
              type: string
          required:
            - id_animal
            - numero_arete
        created_at:
          type:
            - string
            - 'null'
        updated_at:
          type:
            - string
            - 'null'
      required:
        - pesaje_id
        - id_animal
        - peso_kg
        - fecha_pesaje
        - observaciones
        - created_at
        - updated_at
      title: PesajeResource
...
    RazaResource:
      type: object
      properties:
        raza_id:
          type: integer
        nombre:
          type: string
        created_at:
          type:
            - string
            - 'null'
        updated_at:
          type:
            - string
            - 'null'
      required:
        - raza_id
        - nombre
        - created_at
        - updated_at
      title: RazaResource
...
    TratamientoAnimalResource:
      type: object
      properties:
        tratamiento_animal_id:
          type: integer
        id_animal:
          type: integer
        tratamiento_id:
          type: integer
        fecha_aplicacion:
          type:
            - string
            - 'null'
        dosis_ml:
          type: number
        observaciones:
          type:
            - string
            - 'null'
        animal:
          type: object
          properties:
            id_animal:
              type: integer
            numero_arete:
              type: string
          required:
            - id_animal
            - numero_arete
        tratamiento:
          $ref: '#/components/schemas/TratamientoResource'
        created_at:
          type:
            - string
            - 'null'
        updated_at:
          type:
            - string
            - 'null'
      required:
        - tratamiento_animal_id
        - id_animal
        - tratamiento_id
        - fecha_aplicacion
        - dosis_ml
        - observaciones
        - created_at
        - updated_at
      title: TratamientoAnimalResource
    TratamientoResource:
      type: object
      properties:
        tratamiento_id:
          type: integer
        nombre:
          type: string
        descripcion:
          type:
            - string
            - 'null'
        tipo:
          type:
            - string
            - 'null'
        aplicacion:
          type: object
          properties:
            tratamiento_animal_id:
              type: string
            fecha_aplicacion:
              type:
                - string
                - 'null'
            dosis_ml:
              type: string
            observaciones:
              type: string
          required:
            - tratamiento_animal_id
            - fecha_aplicacion
            - dosis_ml
            - observaciones
        created_at:
          type:
            - string
            - 'null'
        updated_at:
          type:
            - string
            - 'null'
      required:
        - tratamiento_id
        - nombre
        - descripcion
        - tipo
        - created_at
        - updated_at
      title: TratamientoResource
...
    UpdateAnimalRequest:
      type: object
      properties:
        numero_arete:
          type: string
          maxLength: 50
        raza_id:
          type: integer
        sexo:
          type: string
          enum:
            - Macho
            - Hembra
        fecha_nacimiento:
          type:
            - string
            - 'null'
          format: date-time
        estado:
          type:
            - string
            - 'null'
          maxLength: 50
        potrero_id:
          type: integer
      title: UpdateAnimalRequest
    UpdateCategoriaRequest:
      type: object
      properties:
        nombre:
          type: string
          maxLength: 100
        tipo:
          type:
            - string
            - 'null'
          maxLength: 50
      title: UpdateCategoriaRequest
    UpdateInsumoRequest:
      type: object
      properties:
        nombre:
          type: string
          maxLength: 150
        categoria_id:
          type: integer
        precio:
          type: number
          minimum: 0
        unidad_medida_id:
          type: integer
      title: UpdateInsumoRequest
    UpdatePesajeRequest:
      type: object
      properties:
        id_animal:
          type: integer
        peso_kg:
          type: number
          minimum: 0
          maximum: 9999.99
        fecha_pesaje:
          type: string
          format: date-time
        observaciones:
          type:
            - string
            - 'null'
      title: UpdatePesajeRequest
    UpdatePotreroRequest:
      type: object
      properties:
        nombre:
          type: string
          maxLength: 100
        hectareas_de_extension:
          type: number
          minimum: 0
          maximum: 999999.99
        capacidad_maxima:
          type: integer
          minimum: 0
        estado_pasto:
          type:
            - string
            - 'null'
          maxLength: 50
      title: UpdatePotreroRequest
    UpdateRazaRequest:
      type: object
      properties:
        nombre:
          type: string
          maxLength: 100
      title: UpdateRazaRequest
    UpdateTratamientoAnimalRequest:
      type: object
      properties:
        id_animal:
          type: integer
        tratamiento_id:
          type: integer
        fecha_aplicacion:
          type: string
          format: date-time
        dosis_ml:
          type: number
          minimum: 0
          maximum: 9999.99
        observaciones:
          type:
            - string
            - 'null'
      title: UpdateTratamientoAnimalRequest
    UpdateTratamientoRequest:
      type: object
      properties:
        nombre:
          type: string
          maxLength: 100
        descripcion:
          type:
            - string
            - 'null'
        tipo:
          type:
            - string
            - 'null'
          maxLength: 50
      title: UpdateTratamientoRequest
    UpdateUnidadMedidaRequest:
      type: object
      properties:
        nombre:
          type: string
          maxLength: 100
        apertura:
          type:
            - string
            - 'null'
          maxLength: 20
      title: UpdateUnidadMedidaRequest
...
  responses:
...
    ValidationException:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Errors overview.
              errors:
                type: object
                description: A detailed description of each field that failed validation.
                additionalProperties:
                  type: array
                  items:
                    type: string
            required:
              - message
              - errors